+++ test/IntegrationTests/AspNetTests.cs
@@ -51,6 +51,72 @@ public class AspNetTests
         collector.AssertExpectations();
     }
 
+    [Theory]
+    [Trait("Category", "EndToEnd")]
+    [Trait("Containers", "Windows")]
+    [InlineData("Classic")]
+    [InlineData("Integrated")]
+    public async Task SubmitTracesCapturesHttpHeaders(string appPoolMode)
+    {
+        Assert.True(EnvironmentTools.IsWindowsAdministrator(), "This test requires Windows Administrator privileges.");
+
+        // Using "*" as host requires Administrator. This is needed to make the mock collector endpoint
+        // accessible to the Windows docker container where the test application is executed by binding
+        // the endpoint to all network interfaces. In order to do that it is necessary to open the port
+        // on the firewall.
+        using var collector = new MockSpansCollector(Output, host: "*");
+        using var fwPort = FirewallHelper.OpenWinPort(collector.Port, Output);
+        collector.Expect("OpenTelemetry.Instrumentation.AspNet.Telemetry", span => // Expect Mvc span
+        {
+            if (appPoolMode == "Classic")
+            {
+                return span.Attributes.Any(x => x.Key == "http.request.header.custom-request-test-header2" && x.Value.StringValue == "Test-Value2")
+                       && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header1")
+                       && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header3")
+                       && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header1")
+                       && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header3")
+                       && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header2");
+            }
+
+            return span.Attributes.Any(x => x.Key == "http.request.header.custom-request-test-header2" && x.Value.StringValue == "Test-Value2")
+                   && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header1")
+                   && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header3")
+                   && span.Attributes.Any(x => x.Key == "http.response.header.custom-response-test-header1" && x.Value.StringValue == "Test-Value4")
+                   && span.Attributes.Any(x => x.Key == "http.response.header.custom-response-test-header3" && x.Value.StringValue == "Test-Value6")
+                   && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header2");
+        });
+
+        collector.Expect("OpenTelemetry.Instrumentation.AspNet.Telemetry", span => // Expect WebApi span
+        {
+            if (appPoolMode == "Classic")
+            {
+                return span.Attributes.Any(x => x.Key == "http.request.header.custom-request-test-header2" && x.Value.StringValue == "Test-Value2")
+                       && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header1")
+                       && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header3")
+                       && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header1")
+                       && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header3")
+                       && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header2");
+            }
+
+            return span.Attributes.Any(x => x.Key == "http.request.header.custom-request-test-header2" && x.Value.StringValue == "Test-Value2")
+                   && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header1")
+                   && span.Attributes.All(x => x.Key != "http.request.header.custom-request-test-header3")
+                   && span.Attributes.Any(x => x.Key == "http.response.header.custom-response-test-header1" && x.Value.StringValue == "Test-Value1")
+                   && span.Attributes.Any(x => x.Key == "http.response.header.custom-response-test-header3" && x.Value.StringValue == "Test-Value3")
+                   && span.Attributes.All(x => x.Key != "http.response.header.custom-response-test-header2");
+        });
+
+        var collectorUrl = $"http://{DockerNetworkHelper.IntegrationTestsGateway}:{collector.Port}";
+        _environmentVariables["OTEL_EXPORTER_OTLP_ENDPOINT"] = collectorUrl;
+        _environmentVariables["OTEL_DOTNET_AUTO_TRACES_ASPNET_INSTRUMENTATION_CAPTURE_REQUEST_HEADERS"] = "Custom-Request-Test-Header2";
+        _environmentVariables["OTEL_DOTNET_AUTO_TRACES_ASPNET_INSTRUMENTATION_CAPTURE_RESPONSE_HEADERS"] = "Custom-Response-Test-Header1,Custom-Response-Test-Header3";
+        var webPort = TcpPortProvider.GetOpenPort();
+        await using var container = await StartContainerAsync(webPort, appPoolMode);
+        await CallTestApplicationEndpoint(webPort);
+
+        collector.AssertExpectations();
+    }
+
     [Fact]
     [Trait("Category", "EndToEnd")]
     [Trait("Containers", "Windows")]
@@ -155,6 +221,10 @@ public class AspNetTests
     private async Task CallTestApplicationEndpoint(int webPort)
     {
         var client = new HttpClient();
+
+        client.DefaultRequestHeaders.Add("Custom-Request-Test-Header1", "Test-Value1");
+        client.DefaultRequestHeaders.Add("Custom-Request-Test-Header2", "Test-Value2");
+        client.DefaultRequestHeaders.Add("Custom-Request-Test-Header3", "Test-Value3");
 
         var response = await client.GetAsync($"http://localhost:{webPort}");
         var content = await response.Content.ReadAsStringAsync();